java-utils-2.eclass: ECLASS_VARIABLE JAVA_PKG_NO_CLEAN
The new ECLASS_VARIABLE JAVA_PKG_NO_CLEAN is defined as an array in
which those *.class or *.jar files can be listed which should not be
removed by java-pkg_clean.  This change helps in cases where only a
small parts of bundled stuff cannot be replaced with system libraries.
It also helps to visualize not yet unbundled stuff in the ebuild.

Signed-off-by: Volkmar W. Pogatzki <gentoo@pogatzki.net>
Signed-off-by: Yuan Liao <liaoyuan@gmail.com>